+++ CHANGELOG.md
@@ -67,6 +67,7 @@
 - support additional redirection op
     - ``<>``: open file descriptor with read-write mode
     - ``<& -``, ``>& -``: close file descriptor
+- truncate large error message of internal error
 
 #### Builtin